Implement a cancelled() signal to fix cancellation of password dialogs
Implement a cancelled() signal in ReadOnlyArchiveInterface. This signal
is emitted in CliInterface when the user cancels a password dialog. The
signal is connected to a slot in Job which sets KJob::error to
KJob::KilledJobError.

The callers (Part and BatchExtract) are modified to do nothing if the
error code is KJob::KilledJobError when running ListJobs/ExtractJobs/
AddJobs/DeleteJobs. This fixes an issue where the user cancels the
"Enter password" dialog and the UI keeps being blocked and/or error
messageboxes appear.

BUG: 267670
FIXED-IN: 15.07.80
REVIEW: 123967