Uncomment out several qCDebug() calls.
There is no point in keeping dead code in the tree. In this case,
however, instead of getting rid of it it makes more sense to just
uncomment out most calls: most of them were commented out during the
port to KF5 in commits such as efe61f2 ("Port to qDebug") and
58b55d1a ("--kdelibs4support, extraction fixes").

A couple of comments were actually removed because they were also
commented out in the KDE4 code.