Implement configuration dialog
This patch implements the standard configuration dialog (using
KConfigDialog) in Ark, as expected from a KDE application.

The widgets showed in the config dialog are provided by Kerfuffle, through the
Ark Part interface. This should help to show the Ark settings in, for
instance, Konqueror's config dialog.

REVIEW: 121997
BUG: 165314
FIXED-IN: 15.08
GUI: