clirartest: Use QSignalSpy to check if entries are correctly parsed
Use QSignalSpy to capture the entry signal emitted from CliPlugin. This
signal is emitted whenever CLiPlugin has found a new archive entry and
contains the ArchiveEntry as argument.

QCOMPARE() is used to check if the correct number of entries were found
and also if the parsed file names were correct.