Add support for overwrite prompt of unrar 5
The overwrite prompt changed in unrar 5 compared to unrar 3 and 4. This
commit adds support for the unrar 5 prompt. Unrar 5 has a multi-line
overwrite prompt, i.e. the last line of the overwrite prompt does not
contain the name of the existing file, which is needed by the
OverwriteQuery to display the filename to the user. This was solved by
storing the filename in a member variable m_storedFileName, so it is
available when the last line of the overwrite prompt is reached. To
capture the filename, a new parameter, FileExistsFileName, was added to
the CliInterfaceParameters enum.

7z also has a multi-line overwrite prompt, so previously the filename
was not displayed in the OverwriteQuery dialog. This is now fixed.

BUG: 349131
FIXED-IN: 15.08.1
REVIEW: 124111