Improve detection of mimetype when opening archive
Previously, the filename extension had higher priority than file content
when determining mimetype. When opening an archive that had a wrong
extension (as in bug 265971), this resulted in Ark failing to open it.
With this commit the mimetype is determined from file content.
Additionally, a warning is shown in console output when mimetype from
extension and content do not match.

For tar.gz, tar.bz2 and tar.xz archives, the mimetype from filename
extension is still used. This is due to these archives being detected as
singlefile gz, bz2 and xz archives when using content, and Ark would
then use the singlefileplugin instead of the libarchive plugin.

BUG: 101170
BUG: 265971
FIXED-IN: 15.08.1
REVIEW: 124442