A fix for drag'n'drop extraction
This is a fix for commit 2072d0d49a8946b62ebb8b3cb961c512ad1f8a25. When
setting the root node by finding the topmost unselected parent,
unselected subfolders of a parent selected folder were not taken into
consideration. This resulted in files in the unselected subfolder having
the wrong root node and hence being extracted into a higher level
folder.

The issue was fixed by excluding unselected folders which are already
part of the index list (due to being subfolders of a selected parent
folder) as root nodes.