Fix drag'n'drop extraction with CLI-plugins
This fixes drag'n'drop extraction of selected entries from Ark to e.g.
Dolphin for archives handled by CLI-plugins. Previously, the entries
would be extracted with full path, which is not what the user expects
and not how the libarchive-plugin handles tar-based archives. This was
due to CLI-plugins not supporting individual root nodes (i.e. removing
a part of the path from entry to be extracted). This is now circumvented
by extracting to a QTemporaryDir, removing the root node from the path,
and finally moving the files to their final destination. The moving to
final destination is done in a new member function in CliInterface
(moveToFinalDest). This function checks if the destination file exists
and prompts the user for action if that is the case.

The finished signal() is now emitted from copyFiles()/addFiles()/
deleteFiles()/list(), instead of from processFinished().

BUG: 208384
FIXED-IN: 15.08.2
REVIEW: 125293