Set InternalID for parent directories that are not in archive
Parent directories of an archive entry do not necessarily have their own
entry in the archive. In this case ArchiveModel creates nodes for all
missing parent directories, but only the EntryMetaDataType::FileName and
EntryMetaDataType::IsDirectory were being set. When extracting and
removing root nodes the EntryMetaDataType::InternalID is used as root
node, so this needs to be set as well.