Add CliInterface::resetParsing()
This function is overridden in the individual CliPlugins and resets
parsing variables, such as ParseState, m_comment, etc. This is needed
because an archive can be listed several times without the plugin being
reloaded, e.g. after adding files. This can result in e.g. comments
being appended to themselves or parsing being started in the wrong
ParseState.
resetParsing() is called in CliInterface::list(), before running the
list process.