Assume that header encryption is disabled
This initialize ReadOnlyArchiveInterface::m_isHeaderEncryptionEnabled to false.

The boolean member ReadOnlyArchiveInterface::m_isHeaderEncryptionEnabled is
only set if the header encryption is explicitly enabled by the user. If not,
its default value is used to enable or disable the header encryption. Since
this variable was not initialized, the behavior was undefined when the checkbox
was not checked.

BUG: 351608
REVIEW: 125372
FIXED-IN: 15.08.2