Fix encryption support for CliPlugins
Currently, support for encryption of entries and/or header is stated in
the desktop file for each plugin by two booleans
(X-KDE-Kerfuffle-EncryptEntries and X-KDE-Kerfuffle-EncryptHeader). This
provides a problem when a plugin supports multiple mimetypes, and the
mimetypes differ in their support for encryption forms. E.g. cli7z
supports both zip and 7z archives, and zip archives only support entry
encryption while 7z archives support both encryption forms. The current
implementation results in the header encryption checkbox in CreateDialog
being enabled for zip archives, which it shouldn't.

This commit changes the two parameters to be QStringLists containing the
mimetypes that support the two forms of encryption. The two properties
are then used in supportedEncryptEntriesMimeTypes() and
supportedEncryptHeaderMimeTypes().

REVIEW: 125096