Set cli7zip as the default plugin for zip archives
This increases the priority of the cli7zip plugin, so that p7zip is now the
default backend used for zip archives handling.
This change was required to address the unicode issues of the clizip plugin,
which resulted in a lot of bugs. While not perfect, p7zip handles better
filenames with non-ASCII characters; in particular it never extracts files
which cannot be opened or not even deleted by Dolphin.

Distributions who ship a patched version of Ark to address the same problem
(e.g. by setting libarchive as default plugin for zips, as Debian/Kubuntu do)
should drop their downstream patches, since they should not be necessary anymore.

BUG: 240727
FIXED-IN: 15.12.0
REVIEW: 125505