Respect the ReadWrite property in desktop file for plugins
Currently, Ark enables add/delete functions in the interface only
according to whether an archive is writable on disk. Plugins can specify
whether they support writing to archives through the
X-KDE-Kerfuffle-ReadWrite property in the desktop file.

This commit stores the ReadWrite property in a boolean member variable in
Archive when loading a plugin. Archive::isReadOnly() is modified to take
the variable into account.

Reviewed in Phabricator.