Add an "Open File" action
The action opens an archive entry with the associated application. This
was previously only possible through the "Open With" action, but this
required several clicks. The "Open With" action is renamed to
"Open File With" for consistency.  Currently, the Preview action is
still the default, i.e. when the user clicks an archive entry the
Preview action is called.

QSignalMapper is used to connect the signals from the Open File, Open
File With and Preview actions to the same slot.

As for Preview, the file is extracted to a QTemporaryDir and then
opened. The temporarily extracted file is monitored for changes using
QFileSystemWatcher and on change a KMessageBox is opened to query the
user if the archive should be updated with the modified file. If the
user accepts slotAddFiles() is called.

The previously unimplemented path argument of slotAddFiles() is used to
allow for updating a file in a subdirectory within the archive.

BUG: 208330
FIXED-IN: 15.12.0
CCBUG: 201162
REVIEW: 125339
GUI: