Do not always expand the first level folders
When there is just one root folder (e.g. typically in archives with source code),
it does make sense to expand the first level folder in the QTreeView widget
(because the user will have to expand it anyway).
In all the other use cases, there is no reason to do that. Instead, it's more
adequate to start with the first level collapsed.

BUG: 303708
FIXED-IN: 15.08.3

Differential Revision: D400