Use ecm_qt_declare_logging_category() cmake macro
This macro was introduced in ECM 5.14.0 and generates logging category
declarations. This removes the necessity of Q_LOGGING_CATEGORY calls in
cpp files. A header with Q_DECLARE_LOGGING_CATEGORY is automatically
generated, so the old logging.h file can be removed.

Also the identifier is now ARK everywhere, so all qCDebug() should be
qCDebug(ARK).

This bumps the required dep version for ECM to 5.14.0.