Fix opening of compressed tar archives with special characters in the extension
Compressed tar archives are detected as singlefile
archives when detecting the mimetype by content.
This means that we cannot rely on it to correctly
open such an archive with special characters in
the extension, as we do for the other plugins.
This patch performs input validation on the
filename extension of compressed tar archives.
Non-alpha characters are removed from the extension,
excluding periods, before detecting the mimetype by name.

BUG: 328815
FIXED-IN: 15.08.3

Differential Revision: D432