Drop KHTML dependency and skip KHTMLPart when previewing
This drops the KHTML framework as dependency (which was actually needed only to
disable javascript and friends from the KHTMLPart) and drops the KHTMLPart as
well, when previewing an HTML or XML file. Now we always use the second available
service offer, which is the katepart by default, to preview those mimetypes.
The user can still render an HTML entry, by opening it in the browser with the
'Open File' action.

BUG: 201162
BUG: 353786
BUG: 354182
FIXED-IN: 15.12.0

Differential Revision: D454