Add support for reading archive comments
Add support for reading comments in CLI-based plugins such as rar and
zip. The comment is parsed from the output of ListJob. The comment is
displayed below the file view in a read-only QPlainTextEdit with
monospace font. A QSplitter is used so the user can resize the comment
field.

Support is added to clirar, clizip and cli7z plugins. The 7z-format
doesn't support comments, but the cli7z-plugin supports reading comments
in e.g. zip archives.

BUG: 328790
FIXED-IN: 15.12.0
REVIEW: 124468