Improve error handling by libarchive plugin
Handle errors when extracting, adding and deleting by emitting the error
signal which makes Part display an error message. Previously, errors
were only output to console.

A new query (ContinueExtractionQuery) was added. This query is used when
an error occurs when extracting an entry to ask the user if the
extraction should be aborted. A checkbox is available to ignore further
errors.

An extra argument was added to copyData() so the filename of the
entry that failed can be output to console.

This commit is partly based on work done by Raphael in RR123832.

Differential Revision: D471