Show correct uncompressed size of huge zipped files
The clizip plugin was converting the entries' original size from string to int,
but that was not enough for huge files whose size does not fit into an int.

We already convert this value to qulonglong in ArchiveModel, so there is no
need to do this conversion also at the plugin level.

BUG: 355969
FIXED-IN: 15.12.0