Do not show a messagebox with empty error message
When the user stops an AddToArchive job (e.g. from the Notifications plasmoid),
the errorText() of the stopped AddJob is empty and a meaningless KMessageBox is
displayed. With this patch, the messagebox is created only if there is an error to show.

As a side effect, Ark stops crashing when using cliplugins, because emitResult() is not
blocked anymore by the messagebox. For the same reason, the end of the cliplugin's
QProcess is not anymore delayed until the user closes the messagebox.

Note that if there would be an error message, Ark would crash anyway.

CCBUG: 222392