Fallback to read-only mode if there are no read-write executables
Commit 2d000a0 introduced executables check when loading a plugin. However the
current behavior is too restrictive: if one wants only to open a rar or a
zip archive, there is no need to require also the rar or zip program to be
installed. Plus, some distributions (e.g. Archlinux) ship only unrar in their
official repositories.

With this commit, Ark is able to understand that e.g. unrar is installed but
rar is not. In this case, Ark can and should fallback to read-only mode,
to disable the Add/Delete actions in the toolbar.

BUG: 357057
FIXED-IN: 15.12.1

CC: rthomsen6@gmail.com