Fix extraction of archives with entries encryption
Differently from other plugins, the unar's PasswordSwitch requires a space
between the switch and the password. Initially we tried to use a "-password
$Password" value for the PasswordSwitch, but this way unar exits with
the error code 1. If we use different strings for "-password" and "$Password"
the extraction works as expected. This also means that $PasswordSwitch needs to be
last in the ExtractArgs, to be sure that the CliInterface substitues correctly
the password argument.