Fix opening of header-encrypted archives
lsar -json does not support interactive password prompts. If an archive is
header-encrypted but no password was given to lsar, its process will exit with
error code 2 and with an error message in stdout. We can match the latter to query
the user for a password, and then we can simply call again lsar using the
password as argument.