Fix broken overwrite queries when fully extracting an archive
unar detects if the stdout has been redirected and blocks the input in such
cases. This prevents Ark from executing the unar's overwrite queries when a file
already exists. Since we already extract to a temporary directory for encrypted
archives, let's just always use a temp dir. Then we move the unpacked files to
the intended destination using Qt functions.