Don't run a ListJob when creating a new archive
Currently if we create a new archive using the New action, Ark creates a
ListJob which is obviously going to fail, since the archive does not exist yet.
Previously Ark would just ignore such failing jobs, but with Ark 15.12 the
clirar plugin actually parses the error message of unrar as a valid comment.

We can solve the problem at the root and only run a ListJob if the archive to
be listed actually exists.

Differential Revision: D712