libarchiveplugin: fix nested xi18n calls
xi18n calls cannot be nested, otherwise the markup gets messed up.
Plugins should use kxi18n instead, but that returns a
KLocalizedString which is not a QString and would require us to change the
ArchiveInterface API.

So an easier solution is to not use xi18n from the libarchive plugin.

Differential Revision: D745

Task: T926