Fix opening of ISO files detected as text/plain.
The workaround in commit b10430b was partly reverted by commit 3906648 and is
now restored by this patch.
Some ISO file may be detected-by-content as text/plain, so this case is not catched
by comparing with application/octect-stream.

The new autotest would have catched the regression, and will prevent it from
happening again.

CCBUG: 354344