Fix archives with single files detected as single folder
If an archive contains only one file, there is no folder in it and
thus is wrong to claim that it is a single folder archive.

If the archive instead contains only one empty folder, we can and should say that it's
single folder.

The unit tests are updated to assert this.