Fix subfolder detection of RPM archives
Filenames in a RPM archive start with the ./ prefix.
This means that "." is detected as the subfolder name by the ListJob.

We now remove this prefix to prevent this meaningless string to show up in the extraction dialog,
as the value of the subfolder QLineEdit. An autotest is added as well to assert
this behavior.

See also https://git.reviewboard.kde.org/r/126756/

CC: rdieter@math.unl.edu