CliInterface: use temp dir only for drag-and-drop extractions
Currently Ark uses a temporary extraction directory when a cliplugin extracts
only some selected files (i.e. when the CliInterface::copyFiles() functions gets a non-empty list of files).

This is wrong. This was meant for drag-and-drop extractions only,
where the fileRootNodePair struct is used (in CliInterface::moveToFinalDest) to remove the root node from the dropped
entries. If the user performs a normal extraction via the extraction dialog,
the current behavior introduce some unintended regressions (see commit
7d1dafa).

The proper way to fix this is to explicitly tell the plugin when we want to
perform a drag-and-drop extraction, in which case the temp dir will be used.

CCBUG: 359628