cliunarchiver: fix opening of header-encrypted archives
If the archive is header-encrypted and the password is wrong,
Ark shows an empty archive, without any warning to the user.

With this patch we check the return code of lsar, which is going to be 1 if the password is
wrong, so that we can warn the user as expected.