Create temporary directories on disk and not in RAM
Currently CliInterface::copyFiles() creates a temporary directory in /tmp,
upon extraction of one or more selected files.

This does not scale well on systems where /tmp is mounted in RAM,
if the RAM is low or if the archive has huge files.

With this patch we create the temporary directory on disk.
We only need to be sure that the temp dir is created on the same disk where the final destination is
(otherwise we write each (possibly huge) file twice).

Differential Revision: D720