Fix archivetest random failures
Due to a race condition, it may happen that the QTemporaryDir object in
CliInterface::copyFiles() is not yet destroyed when the event loop's exec()
in testExtraction() returns. When this happens, the temporary dir used for the
extraction occurs in the test's QDirIterator, causing the QCOMPARE to fail.

This is only a workaround, we should find a better solution.

Task: T1770