Enable The Unarchiver plugin
With this new plugin Ark gets complete support for RAR archives, thanks to the
open source The Unarchiver tool.

This enables distributions to offer RAR support out of the box,
without requiring their users to install the non-free unrar from third-party repositories.

The plugin is heavily unit tested and is designed to be used in a
non-interactive way. This means that the Ark process never has to interact with
the CLI prompts for passwords or overwrite queries issued by the plugin.
To do so, a temporary extraction directory is always used.

FEATURE: 283134
FIXED-IN: 16.04.0