cliunarchiver: fix extraction of empty folders
When we move the extracted files from the temp dir to the final destination,
we were skipping all the directories. Now we skip a directory only if we are
not preserving paths or if the directory is not empty. An empty directory needs
to be moved explicitly, while a non-empty one will be implicitly created by
QDir::mkpath() when moving their children files.

This fixes the related unit test.