Add support for lzipped tar-archives
Support for lzipped tar-archives was enabled in the libarchiveplugin.
Libarchive creates tar.lz archives. However, the lzip format only has a
mimetype for single-compressed file (application/x-lzip) and not for
tar.lz archives. Due to this limitation, a rewrite of CreateDialog was
needed to be able to modify mimetypes in the filter combobox.

BUG: 209418
FIXED-IN: 16.04.0
FEATURE:

Differential Revision: D1098