Drop the karchive plugin
This plugin has been disabled for months and unmantained for years.
It is also functionally superseded by the libarchive plugin.

There was an attempt to fix this plugin (in the karchive-fix branch), but in the end we decided to
definitely drop it.

The KArchive framework is still used by the singlefile plugins, so we can't
drop this dependency as well.