Don't list non-existing archives
This issue was supposed to be fixed by commit e5fcaedc4415,
but actually that patch only addressed the Archive::list() issued by the Part/ArchiveModel.

This patch generalizes the check for the "private" Archive::list() calls as well.

This slightly changes the expected values in the unit test for the properties of
a non-existing archive.

Differential Revision: D1133