Drop support for 7zr/7za executables
Since the cli7zplugin now supports also the zip format, we can't rely anymore on
the 7zr executable (a light version of 7za/7z which can only handle
the 7z format).

Since 7za is often packaged togheter with 7zr (e.g. on *ubuntu systems), we
choose to drop it as well. This will allow us to easily check the availability
of cli7zplugin at runtime, by just looking for a single executable (7z).