Drop RAR entry from Compress service menu
The rar executable (which is proprietary software) is not shipped by any major
distribution, so most users get an error message if they try to use the
RAR entry in the Compress service menu.

Users who do have the rar executable can still choose the 'Compress to...'
entry from the service menu, in order to create rar archives from Dolphin.