Move Delete action to File menu
The Delete action was moved under the Archive menu by commit ea1528c.
However, while it is an action that affects archives, it is also an action
that can be only triggered by selecting a file. That's why it make more sense
to have it under the File menu.

This is also consistent with the right-click context menu, which is now 100%
equal to the File menu.