part: use KMessageWidgets for errors about invalid input
When the user opens (from Dolphin) an archive whose plugin is not available,
the feedback is currently very annoying:

- a KMessageBox pops up with the error message
- the user closes the message box
- an empty (and useless) Ark window shows up

This patch replaces the KMessageBoxes with KMessageWidgets, so that the 3 steps
above are merged into 1.

This patch also move the input validation code from Part::openFile() to a new
Part::isLocalFileValid() method.

Differential Revision: D1167