Improve ExtractionDialog
This enables typing an existing subfolder's name in the location bar and
pressing Return/Enter. Previously when this was attempted, Ark would
ignore what the user typed and extract to baseUrl(). This behavior
deviated from a QFileDialog (which we use for Open Dialog), which simply
enters the existing subfolder.

BUG: 360640
FIXED-IN: 16.04.0

Differential Revision: D1165