Handle "no space left" messages from cliplugins
Previously, Ark didn't handle "no space left" error messages from the
cliplugins. This resulted in either the gui being "busy" continously
(clizip), or that the user didn't get any indication that there was a
problem extracting the archive (cli7z and clirar).

This commit adds a pattern (DiskFullPatterns) for catching these errors.
An errorbox is then displayed to the user. Some of the programs used in
the cliplugins, such as 7z v9.38, doesn't output a specific message,
but then a generic "Extraction failed" message is displayed instead.

BUG: 206383, 206385
FIXED-IN: 16.04.0

Differential Revision: D455