Redesign of CreateDialog
CreateDialog was completely redesigned. The KFileWidget is gone and replaced by
a KUrlRequester for the destination folder and a QLineEdit for the filename.
The encryption options are moved to a KCollapsibleGroupBox and further options
can be added in future KCollapsibleGroupBox's.

AddDialog is also gone and CreateDialog is now used directly instead.