part: connect to KJob::finished instead of result when registering jobs
The Part used to emit the signal ready upon receiving the KJob::result signal,
on jobs completion.

After commit a537c5e, the Part would not receive anymore this signal after
killing a ListJob. According to the KJob docs, this is probably due
to a change in the event loop behavior (due to CliInterface now being in the main
thread) which cause the job to be deleted before than expected.

The finished signal is always guaranteed to be emitted (upon
jobs destruction), so it makes sense to use it instead, in this specific case.

Task: T2225