libarchive: properly stop all jobs
We use the m_abortOperation flag in the same way ListJob does (see commit
acb455da04).

We also need to remove the disconnect() in the Job destructor, which makes Ark
crash if we stop an AddToArchive job. This disconnect() looks like a relic of the
past (see e.g. commit 15fce6f5) and doesn't seem really necessary.

Note that the CliInterface jobs still need to be fixed. So do batch extract
jobs, which are KCompositeJobs and currently ignore the kill button in the
systray.

CCBUG: 222392

Differential Revision: D1466