Ask password when adding files to protected archive
Previously, when opening a non-headerencrypted password protected
archive and adding files to it, the files would be added unprotected.

This commit checks if archive is password-protected and if stored
password is empty. If that's the case, the user is queried for a
password. Currently, it is not possible to check if the entered password
is the same as the one used to create the archive originally, so there
exists a possibility for individual files in the archive having
different passwords. This will be addressed in a separate patch.

Differential Revision: D1508