Find and link Qt5 Concurrent module
Qt5 Concurrent is a separate module, so we need to find it and link to it.
This way we can #include <QtConcurrentRun> without hardcoding the module name.

Most distros include this module in their "basic" qt5 package (e.g. qt5-base on
Archlinux), so this is not really a bug.