Add support for editing/adding archive comments
Support was added for editing comments in supported archive types
(currently only RAR). A new bool was added to plugin json files
("SupportsWriteComment") to indicate support. A new action was added and
is found in Archive menu. The editing of comment is done in the same
QPlainTextEdit used to display comments before. When user modifies
comment, a KMessageWidget pops up with a "Save" button. Actual saving of
comment to archive is achieved by a new job type: CommentJob.

FEATURE: 357594
FIXED-IN: 16.08.0
GUI:
Differential Revision: D1493