Fix plugin loading in plugins tests
KPluginLoader will always find kerfuffle plugins in the system plugin path,
even if QT_PLUGIN_PATH is set to a custom location. This means that in the
plugins test, the loaded plugin will be the system one (because it comes later
in the list of plugins), possibly resulting in an invalid plugin factory and a skipped test.

This patch "gives higher priority" to the plugins found in $QT_PLUGIN_PATH,
ignoring the one in the system path and fixing this issue.