addtoarchive: drop extension only when present
AddToArchive currently tries to drop the extension from any filename,
even if there is no extension in it. We now exploit the mimetype
detected-by-extension: if it's the default mimetype,
then the filename is just a string with some dots but no valid extension.
Otherwise the filename has an actual extension, which we discard.

A testcase is added to reproduce the issue.

BUG: 362690
FIXED-IN: 16.04.1

Differential Revision: D1541