Drop RootNode extraction option
We never set this option and because of this there is a lot of unused code (e.g. all the `RootNodeSwitch` stuff).

Digging in the history, the last time that this option was set was in commit 10bc72f453a253.
This means that this option is fully superseded by `fileNodeRootPair` and we can safely discard all this unused code.

Closes T1942

Differential Revision: D1704