Fix non-resizable infopanel when symlinks have a long target
When a symbolic link points to an entry with a long filename, the infopanel
takes almost all available horizontal space and, even worse, cannot be resized.

This is because there is a single QLabel with a multi-line text, one of which
is the symlink's target. This patch replaces this label with a vertical
layout containing different labels. The targetLabel is now a KSqueezedTextLabel so
that a long target name won't mess up anymore the infopanel size.

BUG: 363719
FIXED-IN: 16.08.0

Differential Revision: D1728