Port service menus to KFileItemAction
The Compress and Extract service menus shown in Dolphin's context menu are now
provided by two new KFileItemAction plugins. This allows us to do a lot of
useful checks before actually showing the actions in the context menu.

We can also stop hardcoding the supported mimetypes for the Extract
menu. This is actually required in order to support the Extract menu when more
than one archive is selected. In this case Dolphin sends to the kfileitemaction
plugins the
generic 'application/octet-stream' mimetype, so we have to manually check
anyway the mimetype of any given URL (to be sure that Ark supports it).

This also paves the way to dynamically generate the service menus (e.g. based
on some kind of user preset).

BUG: 189177
BUG: 268163
FIXED-IN: 16.08.0

CCBUG: 348582

Differential Revision: D1761