Fix opening of files whose mimetype inherits from a supported mimetype
Before 16.04 Ark used to be able to open files whose mimetype is not directly
registered with Ark, but inherits from a mimetype that is supported by Ark.
Examples are .odt and .epub files that both inherits from application/zip.

Since we don't use anymore KMimeTypeTrader (since 16.04), we need to manually
do this check while loading our plugins.

BUG: 363717
FIXED-IN: 16.04.2

Differential Revision: D1725