Add support for LZ4-compressed tar archives
The recently released libarchive 3.2.0 adds support for LZ4-compression.
We utilize this to enable support for LZ4-compressed tarballs in
libarchiveplugin.

Libarchive requires the lz4 executable to be found in path.

cmake < 3.6 is not able to detect version number of libarchive 3.2 and
therefore LZ4-support will not be enabled for older cmake versions, even
though libarchive 3.2 is installed.

Differential Revision: D1821