Reset comment widget when opening an archive
When creating a new archive from an Ark instance where an archive with
comment already is opened, the comment widget from the previous archive
remains visible. This is now fixed by clearing and hiding the
QPlaintTextEdit in beginning of Part::openFile().

Part::slotCommentChanged() gets called when calling
QPlainTextEdit::clear(), so to avoid a segfault we need to check for an
invalid m_model.archive() and in that case return.

Differential Revision: D2114